fix(fileprovider): four macOS-Tahoe correctness bugs surfaced by E2E
1) MapleItem.itemVersion returns non-empty Data for every item shape.
   The protocol default (empty bytes) is rejected by macOS Sequoia/Tahoe
   with `__FILEPROVIDER_BAD_ITEM_MISSING_ITEMVERSION__`, aborting the
   extension mid-enumeration. Library roots and trash containers (which
   carry no mtime) used to hit this; now they hash on the identifier so
   the version is stable across rebuilds.

2) MapleItem(assetMetadata:) and MapleItem(workingSetEntry:) accept an
   explicit `parent:` parameter and stop hardcoding `.workingSet`. The
   call sites resolve the real folder parent from the catalog's library
   roots + absPath — no more "Working Set" container appearing in
   Finder, no more reparent animations when a file's first appearance
   gets relocated by a later folder enumeration.

3) AssetChange.AssetMetadata.mtimeMS is now Double, not Int64. The
   server emits stat.mtimeMs as a float ("1778821543979.7524") and
   Swift's JSONDecoder rejected the fractional part with
   `Number ... is not representable in Swift.`

4) WorkingSetEnumerator now accepts a rootCache and uses the same
   parent-resolution helper. Items routed through working set still
   exist (for OS indexing) but parent under their real folder so they
   never surface as children of a "Working Set" pseudo-folder.

+++ src/apple/Packages/MapleCore/Sources/MapleCore/FileProvider/AssetChange.swift
@@ -134,13 +134,14 @@ public struct AssetMetadata: Codable, Sendable, Equatable {
     public let filename: String
     public let absPath: String
     public let size: Int64
-    /// Epoch MILLISECONDS — convert with `Date(timeIntervalSince1970:
-    /// Double(mtimeMS) / 1000.0)` for SwiftUI `contentModificationDate`.
-    public let mtimeMS: Int64
+    /// Epoch MILLISECONDS. Server emits `stat.mtimeMs` which is a
+    /// floating-point number (sub-millisecond precision); decode as
+    /// `Double` so the JSON parser doesn't reject the fractional part.
+    public let mtimeMS: Double
     public let rating: Int
 
     public init(id: String, folderID: String, filename: String, absPath: String,
-                size: Int64, mtimeMS: Int64, rating: Int) {
+                size: Int64, mtimeMS: Double, rating: Int) {
         self.id = id
         self.folderID = folderID
         self.filename = filename
@@ -153,7 +154,7 @@ public struct AssetMetadata: Codable, Sendable, Equatable {
     /// Convenience: the modification date as a Foundation `Date`,
     /// computed from the server's millisecond timestamp.
     public var contentModificationDate: Date {
-        Date(timeIntervalSince1970: Double(mtimeMS) / 1000.0)
+        Date(timeIntervalSince1970: mtimeMS / 1000.0)
     }
 
     private enum CodingKeys: String, CodingKey {
